+++ src/view/task/task.ts
@@ -1,75 +1,45 @@
-import {
-  getFormattedDate,
-  checkIsTaskExpired,
-  checkIsTaskRepeating,
-} from '~/helpers';
+import { createElement } from '~/helpers';
 import { ITask } from '~/common/interfaces';
-import { DateFormatType } from '~/common/enums';
+import { createTaskTemplate } from './templates/task-template/task-template';
+import { EMPTY_TASK, TaskTemplateMode } from './common';
+import { createTaskEditTemplate } from './templates/task-edit-template/task-edit-template';
 
-const createTaskTemplate = (task: ITask) => {
-  const {
-    color,
-    description,
-    dueDate,
-    repeating,
-    isArchive,
-    isFavorite,
-  } = task;
+class Task {
+  #element: Element | null;
 
-  const date = dueDate
-    ? getFormattedDate(DateFormatType.FULLMONTH_DAY, dueDate)
-    : ``;
+  #task: ITask;
 
-  const deadlineClassName = checkIsTaskExpired(dueDate) ? `card--deadline` : ``;
-  const repeatClassName = checkIsTaskRepeating(repeating) ? `card--repeat` : ``;
-  const archiveClassName = isArchive
-    ? `card__btn--archive card__btn--disabled`
-    : `card__btn--archive`;
+  #templateMode: TaskTemplateMode;
 
-  const favoriteClassName = isFavorite
-    ? `card__btn--favorites card__btn--disabled`
-    : `card__btn--favorites`;
+  constructor(task: ITask | null) {
+    this.#task = task ?? EMPTY_TASK;
+    this.#templateMode = task ? TaskTemplateMode.DEFAULT : TaskTemplateMode.EDIT;
+    this.#element = null;
+  }
 
-  return `
-  <article class="card card--${color} ${deadlineClassName} ${repeatClassName}">
-    <div class="card__form">
-      <div class="card__inner">
-        <div class="card__control">
-          <button type="button" class="card__btn card__btn--edit">
-            edit
-          </button>
-          <button type="button" class="card__btn ${archiveClassName}">
-            archive
-          </button>
-          <button
-            type="button"
-            class="card__btn ${favoriteClassName}"
-          >
-            favorites
-          </button>
-        </div>
-        <div class="card__color-bar">
-          <svg class="card__color-bar-wave" width="100%" height="10">
-            <use xlink:href="#wave"></use>
-          </svg>
-        </div>
-        <div class="card__textarea-wrap">
-          <p class="card__text">${description}</p>
-        </div>
-        <div class="card__settings">
-          <div class="card__details">
-            <div class="card__dates">
-              <div class="card__date-deadline">
-                <p class="card__input-deadline-wrap">
-                  <span class="card__date">${date}</span>
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </article>`;
-};
+  get node() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
 
-export { createTaskTemplate };
+    return this.#element;
+  }
+
+  get template() {
+    // eslint-disable-next-line default-case
+    switch (this.#templateMode) {
+      case TaskTemplateMode.DEFAULT:
+        return createTaskTemplate(this.#task);
+      case TaskTemplateMode.EDIT:
+        return createTaskEditTemplate(this.#task);
+    }
+
+    return null;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
+
+export default Task;
