+++ src/view/task/templates/task-edit-template/templates/task-color/task-color.ts
